Fixed inconsistancy
Noticed inconsistancys on graphs as I used the average of the score column on most except 2 which I used the av_score column which was not accurate. So i added a new column to the dataset score_avg which showed the true average of the score column for each item and updated the plots with using that column.